+++ fclneuron.cpp
@@ -8,22 +8,23 @@ FCLNeuron::FCLNeuron(const int& i):index(i)
 
 }
 
-std::pair<double, double> FCLNeuron::setActivation(const double& sum)
+std::pair<long double, long double> FCLNeuron::setActivation(
+        const long double& sum)
 {
     //use sigmoid
-    double sigmoid = 1/(1 + exp(-sum));
+    long double sigmoid = 1/(1 + exp(-sum));
     activation.first = sigmoid;
     activation.second = sigmoid * (1 - sigmoid);
     inputSum = sum;
     return activation;
 }
 
-std::pair<double, double> FCLNeuron::getActivation() const
+std::pair<long double, long double> FCLNeuron::getActivation() const
 {
     return activation;
 }
 
-double FCLNeuron::getInput() const
+long double FCLNeuron::getInput() const
 {
     return inputSum;
 }
